+++ console/src/components/Services/Data/TableModify/ColumnEditor.js
@@ -11,7 +11,7 @@ import {
   JSONB,
   TIMESTAMP,
   TIME,
-} from '../../../../constants';
+} from '../utils';
 
 const ColumnEditor = ({
   column,
